Move location client implementations
