+++ collect_app/src/test/java/org/odk/collect/android/location/client/GoogleFusedLocationClientTest.java
@@ -16,6 +16,7 @@ import org.junit.runner.RunWith;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 import org.odk.collect.android.location.LocationTestUtils;
+import org.odk.collect.location.GoogleFusedLocationClient;
 import org.odk.collect.location.LocationClient;
 
 import static android.location.LocationManager.GPS_PROVIDER;
