+++ location/build.gradle
@@ -35,5 +35,6 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.6.0'
     implementation "com.google.android.gms:play-services-location:17.1.0"
+    implementation "com.jakewharton.timber:timber:${rootProject.timberVersion}"
     testImplementation 'junit:junit:4.13.2'
 }
