+++ location/src/main/java/org/odk/collect/location/AndroidLocationClient.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.location.client;
+package org.odk.collect.location;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -6,12 +6,10 @@ import android.location.Location;
 import android.location.LocationManager;
 import android.os.Bundle;
 
-import com.google.android.gms.location.LocationListener;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import org.odk.collect.android.utilities.GeoUtils;
+import com.google.android.gms.location.LocationListener;
 
 import timber.log.Timber;
 
@@ -25,7 +23,7 @@ import timber.log.Timber;
  * Package-private, use {@link LocationClientProvider} to retrieve the correct
  * {@link LocationClient}.
  */
-class AndroidLocationClient
+public class AndroidLocationClient
         extends BaseLocationClient
         implements android.location.LocationListener {
 
@@ -40,7 +38,7 @@ class AndroidLocationClient
      *
      * @param context The Context where the AndroidLocationClient will be running.
      */
-    AndroidLocationClient(@NonNull Context context) {
+    public AndroidLocationClient(@NonNull Context context) {
         this((LocationManager) context.getSystemService(Context.LOCATION_SERVICE));
     }
 
@@ -50,7 +48,7 @@ class AndroidLocationClient
      *
      * @param locationManager The LocationManager to retrieve locations from.
      */
-    AndroidLocationClient(@NonNull LocationManager locationManager) {
+    public AndroidLocationClient(@NonNull LocationManager locationManager) {
         super(locationManager);
     }
 
@@ -113,7 +111,7 @@ class AndroidLocationClient
     public Location getLastLocation() {
         String provider = getProvider();
         if (provider != null) {
-            return GeoUtils.sanitizeAccuracy(getLocationManager().getLastKnownLocation(provider));
+            return LocationUtils.sanitizeAccuracy(getLocationManager().getLastKnownLocation(provider));
         }
 
         return null;
@@ -148,7 +146,7 @@ class AndroidLocationClient
         Timber.i("Location changed: %s", location.toString());
 
         if (locationListener != null) {
-            locationListener.onLocationChanged(GeoUtils.sanitizeAccuracy(location));
+            locationListener.onLocationChanged(LocationUtils.sanitizeAccuracy(location));
         }
     }
